Add files via upload
This is a simple code for an alarm system that uses a RPI to text/email you if a motion detector has been triggered. It arms and disarms with an MFRC module/prox card. After arming there will be a beeping noise and a delay of 90 seconds to leave and a blinking green light. Upon setting off the alarm you have 90 seconds to swipe your prox card while a red light blinks and the buzzer beeps at you. If you swipe your card in time the light will turn steady green and the noise will stop. If you do not it will email or text you that your alarm has been triggered. Simple yet effective.